test: switch tests to jest
