+++ package.json
@@ -5,11 +5,12 @@
 	"main": "dist/index.js",
 	"types": "dist/index.d.ts",
 	"scripts": {
-		"lint": "eslint src --ext .ts",
+		"lint": "eslint test src --ext .ts",
+		"lint:fix": "eslint test src --ext .ts --fix",
 		"prebuild": "npm run lint",
 		"build": "rimraf dist/ && tsc",
 		"pretest": "npm run build",
-		"test": "node test/index.js",
+		"test": "jest --verbose",
 		"docs": "docgen --jsdoc jsdoc.json --source src/*.ts src/**/*.ts --custom docs/index.yml --output docs/docs.json",
 		"docs:test": "docgen --jsdoc jsdoc.json --source src/*.ts src/**/*.ts --custom docs/index.yml"
 	},
@@ -51,6 +52,7 @@
 		"lint-staged": "^10.5.3",
 		"prettier": "^2.2.1",
 		"rimraf": "^3.0.2",
+		"ts-jest": "^26.4.4",
 		"typescript": "^4.1.3"
 	},
 	"eslintConfig": {
@@ -114,5 +116,9 @@
 				]
 			]
 		}
+	},
+	"dependencies": {
+		"@types/jest": "^26.0.20",
+		"jest": "^26.6.3"
 	}
 }
