+++ tsconfig.jest.json
@@ -1,0 +1,5 @@
+{
+	"extends": "./tsconfig.json",
+	"module": "commonjs",
+	"include": ["test"]
+}
